refactor: Complete rewrite of the library
