+++ src/stringify.ts
@@ -1,9 +1,15 @@
-import type { SomeSchema } from '../types';
+import type { SomeSchema } from './types';
+
+export const schemaToString = (schema: SomeSchema<any>): string => {
+  const modifiers = getModifiers(schema);
+  return unionToPrint([schema.toString(), ...modifiers]);
+};
 
 export const typeToPrint = (str: string) => str;
 export const objectToPrint = (str: string) => '{' + str + '}';
 export const quote = (str: string) => (/\s/.test(str) ? `"${str}"` : str);
-export const unionToPrint = (arr: readonly string[]): string => {
+
+const unionToPrint = (arr: readonly string[]): string => {
   const str = arr.join(' | ');
 
   if (arr.length > 1) {
@@ -11,7 +17,8 @@ export const unionToPrint = (arr: readonly string[]): string => {
   }
   return str;
 };
-export const getModifiers = (v: SomeSchema<any>): readonly string[] => {
+
+const getModifiers = (v: SomeSchema<any>): readonly string[] => {
   const modifiers = [v.__modifiers.optional && 'undefined', v.__modifiers.nullable && 'null'];
   return modifiers.filter((m): m is string => Boolean(m));
 };
