+++ src/validators/boolean.ts
@@ -1,9 +1,8 @@
 /* eslint-disable functional/no-this-expression */
 import { ValidationError } from '../errors';
+import { initialModifiers } from '../schema';
+import { typeToPrint } from '../stringify';
 import type { Schema } from '../types';
-
-import { initialModifiers } from './__schema';
-import { typeToPrint } from './__stringifyHelpers';
 
 export const boolean = () => {
   return {
